requirements.txt
# LangChain Core
langchain
langchain-Core
langchain-community

# OpenAI Integration
langchain-openai
openai
transformers
sentence-transformers


# Vector Store
faiss-cpu 

# Environment Variable Management
python-dotenv

#UI
streamlit

# Machine learning Utilities
numpy
scikit-learn
tqdm